+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/dccticketing/ui/consent/one/DccTicketingConsentOneViewModel.kt
@@ -5,13 +5,13 @@ import androidx.lifecycle.asLiveData
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
+import de.rki.coronawarnapp.R
 import de.rki.coronawarnapp.dccticketing.core.common.DccTicketingException
-import de.rki.coronawarnapp.dccticketing.core.service.DccTicketingRequestService
 import de.rki.coronawarnapp.dccticketing.core.transaction.DccTicketingTransactionContext
 import de.rki.coronawarnapp.dccticketing.ui.shared.DccTicketingSharedViewModel
 import de.rki.coronawarnapp.util.coroutine.DispatcherProvider
-import de.rki.coronawarnapp.util.encryption.ec.EcKeyGenerator
 import de.rki.coronawarnapp.util.ui.SingleLiveEvent
+import de.rki.coronawarnapp.util.ui.toResolvingString
 import de.rki.coronawarnapp.util.viewmodel.CWAViewModel
 import de.rki.coronawarnapp.util.viewmodel.CWAViewModelFactory
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -22,8 +22,7 @@ import timber.log.Timber
 class DccTicketingConsentOneViewModel @AssistedInject constructor(
     @Assisted private val dccTicketingSharedViewModel: DccTicketingSharedViewModel,
     dispatcherProvider: DispatcherProvider,
-    private val dccTicketingRequestService: DccTicketingRequestService,
-    private val ecKeyGenerator: EcKeyGenerator
+    private val dccTicketingConsentOneProcessor: DccTicketingConsentOneProcessor
 ) : CWAViewModel(dispatcherProvider = dispatcherProvider) {
 
     private val currentIsLoading = MutableStateFlow(false)
@@ -48,82 +47,27 @@ class DccTicketingConsentOneViewModel @AssistedInject constructor(
     fun onUserConsent(): Unit = launch {
         Timber.d("onUserConsent()")
         currentIsLoading.compareAndSet(expect = false, update = true)
-        processUserConsent()
+        val event = try {
+            Timber.d("processUserConsent()")
+            dccTicketingSharedViewModel.apply {
+                val ctx = transactionContext.first()
+                dccTicketingConsentOneProcessor.processUserConsent(ctx = ctx)
+                    .also { updateTransactionContext(ctx = it) }
+            }
+            NavigateToCertificateSelection
+        } catch (e: Exception) {
+            Timber.e(e, "Error while processing user consent")
+            val lazyErrorMessage = when (e) {
+                is DccTicketingException -> {
+                    val serviceProvider = currentUiState.first().provider
+                    e.errorMessage(serviceProvider = serviceProvider)
+                }
+                else -> R.string.errors_generic_text_unknown_error_cause.toResolvingString()
+            }
+            ShowErrorDialog(lazyErrorMessage = lazyErrorMessage)
+        }
+        postEvent(event = event)
         currentIsLoading.compareAndSet(expect = true, update = false)
-    }
-
-    private suspend fun processUserConsent() = try {
-        Timber.d("processUserConsent()")
-        dccTicketingSharedViewModel.apply {
-            transactionContext.first()
-                .requestServiceIdentityDocumentOfValidationService()
-                .generateECKeyPair()
-                .requestAccessToken()
-                .also { updateTransactionContext(ctx = it) }
-        }
-        postEvent(NavigateToCertificateSelection)
-    } catch (e: DccTicketingException) {
-        Timber.d(e, "Error while processing user consent")
-        val serviceProvider = currentUiState.first().provider
-        val lazyErrorMessage = e.errorMessage(serviceProvider = serviceProvider)
-        postEvent(ShowErrorDialog(lazyErrorMessage = lazyErrorMessage))
-    }
-
-    private suspend fun DccTicketingTransactionContext.requestServiceIdentityDocumentOfValidationService():
-        DccTicketingTransactionContext {
-            Timber.d("requestServiceIdentityDocumentOfValidationService")
-
-            requireNotNull(validationService) { "ctx.validationService must not be null" }
-            requireNotNull(validationServiceJwkSet) { "ctx.validationServiceJwkSet must not be null" }
-
-            val document = dccTicketingRequestService
-                .requestValidationService(validationService, validationServiceJwkSet)
-
-            return copy(
-                validationServiceEncKeyJwkSetForRSAOAEPWithSHA256AESCBC =
-                document.validationServiceEncKeyJwkSetForRSAOAEPWithSHA256AESCBC,
-                validationServiceEncKeyJwkSetForRSAOAEPWithSHA256AESGCM =
-                document.validationServiceEncKeyJwkSetForRSAOAEPWithSHA256AESGCM,
-                validationServiceSignKeyJwkSet =
-                document.validationServiceSignKeyJwkSet
-            )
-        }
-
-    private fun DccTicketingTransactionContext.generateECKeyPair(): DccTicketingTransactionContext {
-        Timber.d("generateECKeyPair()")
-
-        val ecKeyPair = ecKeyGenerator.generateECKeyPair()
-
-        return copy(
-            ecPublicKey = ecKeyPair.publicKey,
-            ecPrivateKey = ecKeyPair.privateKey,
-            ecPublicKeyBase64 = ecKeyPair.publicKeyBase64
-        )
-    }
-
-    private suspend fun DccTicketingTransactionContext.requestAccessToken(): DccTicketingTransactionContext {
-        Timber.d("requestAccessToken()")
-
-        requireNotNull(accessTokenService) { "ctx.accessTokenService must not be null" }
-        requireNotNull(accessTokenServiceJwkSet) { "ctx.accessTokenServiceJwkSet must not be null" }
-        requireNotNull(accessTokenSignJwkSet) { "ctx.accessTokenSignJwkSet must not be null" }
-        requireNotNull(validationService) { "ctx.validationService must not be null" }
-        requireNotNull(ecPublicKeyBase64) { "ctx.ecPublicKeyBase64 must not be null" }
-
-        val accessToken = dccTicketingRequestService.requestAccessToken(
-            accessTokenService = accessTokenService,
-            accessTokenServiceJwkSet = accessTokenServiceJwkSet,
-            accessTokenSignJwkSet = accessTokenSignJwkSet,
-            authorization = initializationData.token,
-            validationService = validationService,
-            publicKeyBase64 = ecPublicKeyBase64
-        )
-
-        return copy(
-            accessToken = accessToken.accessToken,
-            accessTokenPayload = accessToken.accessTokenPayload,
-            nonceBase64 = accessToken.nonceBase64
-        )
     }
 
     private fun postEvent(event: DccTicketingConsentOneEvent) {
